Allow pinning of the git resource ref
This allows users to specify the version of the git repository that should be
checked out rather than following master. It is designed to be used in
situations where a third-party git repository is part of a pipeline and you
wish to have control over the version available, or where your pipeline
determines the required version of a repository at runtime.

The `check` action gains a `pin_ref` parameter that expects a valid git
reference, and will cause the resource to remain pinned at that reference. This
should be used when you wish to lock build dependencies from your pipeline
configuration, such as the version of cf-release.

The `in` action gains a `pin_ref_file` parameter that expects a path to a file
containing a valid git reference, and will cause the resource to remain pinned
at that reference. This should be used when the version of the repository
required is determined by a step earlier in the pipeline.