test get returns the tag name if tag_filter is set
When using tag_filter, the check step will return a list of valid
tags as versions. In order to keep consistency, the get command
should also return the tag if the tag_filter is defined.

In this test we verify that we return the latest tag matching tag_filter
that points to $ref, regardless $ref is a tag, a hash or if it matches
or not tag_filter. If no tags match it, we will return the hash pointer
to $ref.

Tests #45