get returns the tag name if tag_filter is set
The in command will check if tag_filter is defined in the source
configuration, and if so, will return the matching tag as a version
or the hash if none match.

In this implementation if several tags match, we order the tags by version
number and return the latest one.

Fixes #45