use show-ref to get branch names for metadata
Using rev-parse of head will always return the commit hash when
working in a detached HEAD, even if the HEAD ref is being also
referenced by the branch. The in command works detached
whenever a ref is passed to the in command.

The alternative is list all the heads and filter the ones pointing to
the current commit.