Add GPG configuration fixtures
Keys can't be generated from inside the container as there is not
enough entropy. Instead test keys have been pregenerated so they
can be imported into GPG.

Signing things with gpg without requiring passphrase input is tricky,
and very specific to the gpg version in use (gpg 2.0 is quite different
to gpg 2.1 for example), so this includes a fixture repo with a head
commit that's signed by this private key.