Disable glob expansion when checking paths
Update the `/check` script to disable bash glob expansion of path
patterns.

As of Git 1.8.4 the platform fnmatch(3) function is used for pathspec
matching allowing patterns like "**/file" to match both "a/b/file" and
"a/foo". Unfortunately it's also possible for bash to expand these
patterns which can lead to unexpected results.

For example, prior to this commit, given files added to git in the
following order:

- 'a/file'
- 'a/b/file'

Calling the `/check` script with the pattern `**/file` would return
the commit commit containing `a/file` instead of the expected later
`/a/b/file` commit.

Fixes gh-120