Add a signed tag fixture repo
Add a signed tag fixture to use for testing the verification of tags.
The gpg verification tests were using a fixture repository to avoid
having to generate temporary keys, which fails due the lack of entropy.

We want to use the same strategy for testing the verification of tags.
Using a different repository would make all the testing more easy.

We extend the generate_fixture_repo to create the sequencially.
We use the config -c user.signingkey='...' to pass the right user key,
as signing tags do not allow pass the key as argument.

I had to pass the variable GPG_TTY when using pinentry-ncurses with git tag.