Allow specify verification of commits or tags
Allow pass the parameter `verify` with a list of elements to verify.
Valid values are "commit" and "tags".

"commit": Will verify the current commit signed, meanwhile "tags"
will verify that AT LEAST one tag is signed with the provided keys.

The default value is `verify: []`. if the user passes `verify: []` it
would disable any the verification.