Populate .git/ref with the reference checked out
When using `tag_filter`, the version reported by `check` would be
the matching tag name.

But if multiple tags point to the same commit, it is not possible to know
which tag in the repository is being used to checkout the code in
concourse.

By writing the value of `$ref` in `.git/ref` the consumers of the
resource can know the exact version used by concourse, regarless it is
a tag or a commit SHA-1.