Make shallow cloning more durable
If a git-resource is configured to use shallow cloning and new commits
are added to the repository that are not considered new versions of the
resource (e.g. via a path filter excluding it), a `get` will shallow
clone HEAD, which will not be equal to the version ref, and start to
break when it then tries to checkout the version ref since there is
no longer sufficient history available to find version ref. One
solution, which is proposed here, is to always fetch enough depth to
checkout the version ref corresponding to that version of the resource.