Fix annotate file path
Because the annotation file's path is relative and because of the `cd
$repository` that occurs in `assets/out`, if you want to use an upstream
output directory for the contents of the tag and the annotate files, you
have to do something like the following:

```yaml
- put: some-git-repo
  params:
    repository: some-git-repo
    tag: tag/tag
    annotate: ../tag/annotate
```

This is surprising and probably not what was intended. You'd rather
specify it like this:

```yaml
- put: some-git-repo
  params:
    repository: some-git-repo
    tag: tag/tag
    annotate: tag/annotate
```

The test for this was fixed as well. In the normal case you cannot pass
the source directory as part of the path; you can only pass a path
relative to the source directory.