in script should not retain tags
We were encountering a problem very similar to #30: when we manually
deleted tags from our repo, our Concourse pipeline was then re-pushing
them. We figured out that since Concourse had a cached version of our
repo with the tags in it, and since it pushes with tags every time, this
caused it to recreate the tags.