Invoke `git-crypt` directly to avoid absolute paths in .git/config
When git-crypt is invoked as `git crypt`, git calls git-crypt using an
absolute path. Therefore the filter and diff config that git-crypt puts
into .git/config will contain the absolute path to git-crypt. This is a
problem for tasks that perform commits, as the git-crypt path in the
task container may not match the one in the resource container. This
allows task images to have git-crypt wherever they like in $PATH.

Signed-off-by: Giuseppe Capizzi <gcapizzi@pivotal.io>