Improve documentation for `depth` parameter
Now that the `git` resource will deepen a shallow clone enough times to
ensure it has `version.ref` in its history, we can remove the warranty
claim. Because this is scenario feels non-obvious, I've also added a
succint description of the problem and how it will be resolved by the
resource.