Fixes assertions in submodule parameter test.
These assertions weren't testing the behaviour that differed around the
submodule_recursive flag, which is whther nested submodules get checked
out. The assertions seemed to be more about the depth flag (and the
number of commits present). This updates them to assert on the number of
files present in the various subdirs.

This also updates the assertion for the checked out commit to test the
destination dir instead of the source.