check: Return commit from the tag as resource version
Current git resource is only watch tags name but not tags update. So
pipelines aren't triggered when the tag is updated and the resource

with `error: pathspec 'DEV' did not match any file(s) known to git.`
To solve it I suggest to not use only ref but ref and commit when
tag_filter is used.
This commit allow to store the tag name and commit id associeted to

tag as "version" of the resource. With thos way we are able to
re-trigger the pipeline when a the tags is updated.