Properly support boolean values for submodule_recursive.
The previous implementation didn't work correctly if the value was
specified as a boolean false, becuase the logic would then fall through
to the next option, ending at the default value of true.

This instead constructs an array of the possible values and selects the
first non-null one, which works correctly.

Signed-off-by: Alex Tomlins <alex.tomlins@digital.cabinet-office.gov.uk>