refactor/simplify submodule deepening
don't bother restoring config; it should be empty, as the repo was
freshly cloned. just unset it after so the implementation detail doesn't
leak to builds.

moved init to --init on the update command, and made it so everything
looks at .gitmodules, rather than relying on 'git submodule init' to
populate repo config.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>