Support more commit URL possibilities in metadata
The existing logic in `add_git_metadata_url` tries to synthesize a URL
of the form "https://github.com/org/repo/commit/abc123", in the case
where we know we're working with a GitHub repo. The URL can be
displayed in the Concourse UI, so users can click to navigate straight
to the commit in question.

Issue #279 reported that this did not work correctly for a given
GitHub Enterprise remote URL containing the substring "github.com",
because the code just looked for that string, and tried to apply some
straightforward rewrite rules in order to generate the commit URL.

More generally, it would be nice to be able to support these URLs for
a wider range of hosting options.

In this change, we switch to parsing out various components of the
remote URL, so that the target hostname can be identified more
reliably; based on that hostname, we guess a commit URL format. The
guessing logic should be enough to identify repos hosted on GitHub,
GitLab or BitBucket shared hosting, plus self-hosted instances of
any of them (or of gogs.io) in the case where the hostname includes
"github", "gitlab", etc.

This therefore _doesn't_ identify a GitHub Enterprise instance living
at something like "code.somebigcompany.com"; there's no way to tell
this is GHE without either being told explicitly, or throwing traffic
at it in order to guess. Neither of those options seem really
attractive for the time being.

Additionally, there are a lot of ways that the source URI _might_ be
set up. Only some of these are commonly used with the providers we're
targeting here. This does support the following:

* https://github.com/org/repo.git
* ssh://git@github.com/org/repo.git
* ssh://git@github.com:443/org/repo.git (for if ssh is blocked)
* git@github.com:org/repo.git

and their non-GitHub parallels, which seem to represent the most
typical configurations.

Signed-off-by: Alexander Gurney <alexander_gurney@cable.comcast.com>