B: allow configuring source.ref to track named ref
this commit adds `source.ref`, which is used to track a named ref in a
git commit. it is intended as an alternative to `source.branch` when the
thing you want to track isn't a branch.

for instance, you can track GitHub pull requests using the ref name
`pull/ID/head`

check and get both fetch the single ref by:
1. initializing an empty git repo
2. adding the uri as a remote
3. fetching the ref

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>