Add search_remote_refs source parameter.
This parameter compliments the `refs_prefix` and enables the round trip
for gerrit's refs/for/<branch> change review workflow.

A gerrit user creates a change for review by pushing to
`refs/for/<branch>`, but then the gerrit remote will rename that ref to
`refs/changes/xxx/yy`.

Since we don't know the actual ref that will be created, we must search
the remote for a ref containing the version we just pushed, and fetch
from there.

Also need to remember to have a separate input and output resource so
that the pushed version isn't counted as an input in a later run.

Signed-off-by: Jamie Pate <jpate@fortinet.com>