workaround checkout-error for nested shallow submodules
If a nested submodule is configured to be "shallow", referencing a
commit that is not reachable from the respective repository's HEAD
(typically the default branch, if e.g. hosted on GitHub), will result in
an error when checking out the submodule (git submodule update..).

Workaround this by always setting `--no-recommend-shallow`, which
lets git ignore this setting.