check: use bare clone
The check container doesn't make any use of an actual checkout, but on
large repositories the time to checkout after fetching can be 90% or
more of the overall time to clone.

Because the --bare option prevents the --branch option from being
setting the refspec in the config we have to do it as a follow-up step.

Instead of relying solely on the config having the refspec set
correctly, we can also explicitly specify the branch when fetching if
one is set.

Signed-off-by: Ryann Graham <r.m.graham@gmail.com>